Reprieve for original HardPencil and SoftPencil commands
PencilHard and PencilSoft benefited from autocomplete when you didn't know the commands, but they are awkward when you do know the commands.

So, there's PencilHard and HardPencil.
PencilSoft and SoftPencil
PencilToggle and TogglePencil
PencilOff and NoPencil

Additional blacklist adjustment for the plasticboy plugin, though I'll be doing more work in the coming weeks to get things tuned better.