Add -bar suffix to commands to allow use as first of series of commands with |
For example, `Pencil|Goyo`. Commands which can be followed by a `|` must be declared with `command-bar` (not entirely clear to me why).

My use case is in making an alias for a pager for prose: `vim -c 'Pencil|Goyo'`, but it could be potentially useful for a variety of reasons.